+++ tests/Message_Queue_Notifications_Test.cpp
@@ -205,7 +205,7 @@ Watermark_Test::producer (void)
 {
   int i = watermark_iterations;
 
-  for (ssize_t hwm = this->hwm_;
+  for (int hwm = this->hwm_;
        hwm >= 0 ;
        hwm -= this->len_)
     {
