+++ tests/Priority_Buffer_Test.cpp
@@ -68,7 +68,7 @@ consumer (void *args)
 
       local_count++;
 
-      size_t length = mb->length ();
+      int length = mb->length ();
 
       if (length > 0)
         {
